Graded HW2
Nicely done! Your functions are clear and readable, and I love how clean your code was in general.
You need to resolve merge conflicts as they occur. When git recognizes a merge conflict, it will insert the conflict code into your code, which then needs to be addressed.

Close this pull request to earn +1 on this homework assignment. Don't merge!
**3.75/5**